feat(quota-pilot): 新增真实额度查询与供应商配置

+++ src/minimaxresponseparser.h
@@ -10,8 +10,8 @@ struct MiniMaxPlan
 {
     QString planId;
     QString planName;
-    int used = 0;
-    int total = 100;
+    double used = 0;
+    double total = 100;
     qint64 resetAtMs = 0;
 };
